# Preview image caching