--- a/Gemfile
+++ b/Gemfile
@@ -20,8 +20,8 @@ gem "redis", "~> 5.1"
 # gem 'bcrypt', '~> 3.1.7'
 
 # Use Active Storage variant
-gem 'image_processing', '~> 1.2'
-gem 'rubyzip'
+gem "image_processing", "~> 1.2"
+gem "rubyzip"
 gem "dotenv-rails", "~> 3.0"
 gem "acts-as-taggable-on", "~> 10.0"