--- a/app/controllers/images_controller.rb
+++ b/app/controllers/images_controller.rb
@@ -6,8 +6,6 @@ class ImagesController < ApplicationController
     image_service = ImageService.new
     resized_image_path = image_service.get_resized_image(original_image_path)
 
-    send_file resized_image_path, type: 'image/jpeg', disposition: 'inline'
+    send_file resized_image_path, type: "image/jpeg", disposition: "inline"
   end
 end
-
-