--- a/app/views/models/_file.html.erb
+++ b/app/views/models/_file.html.erb
@@ -1,8 +1,10 @@
 <div class="col mb-4">
   <div class="card preview-card <%= (file === @model.preview_file) ? "border-primary" : "" %>">
     <% if file.is_image? %>
-      <%= content_tag :div, nil, class: "card-img-top card-img-top-background", style: "background-image: url(#{library_model_model_file_path(@library, @model, file, format: file.extension)})" %>
-      <%= image_tag library_model_model_file_path(@library, @model, file, format: file.extension), class: "card-img-top image-preview", alt: file.name %>
+      <% original_image_path = file.pathname %>
+      <% resized_image_url = resized_image_path(original_image_path: original_image_path) %>
+      <%= content_tag :div, nil, class: "card-img-top card-img-top-background", style: "background-image: url(#{resized_image_url})" %>
+      <%= image_tag resized_image_url, class: "card-img-top image-preview", alt: file.name %>
     <% elsif renderable?(file.extension) %>
       <div class="card-img-top">
         <%= render partial: "object_preview", locals: {library: @library, model: @model, file: file} %>