## Checklist

🚨 Please review the [guidelines for contributing](../CONTRIBUTING.md) to this repository. 🚨

- [x] Make sure you are making a pull request against our **main branch** (left side)
- [x] Check that that your branch is up to date with our main.
- [x] Make sure you are requesting to **pull a topic/feature/bugfix branch** (right side). Don't request *your* main!
- [x] Check that the tests and code linter both pass.
- [x] If you're a new contributor, please sign our [contributor license agreement](https://cla-assistant.io/manyfold3d/manyfold).

## Warnings

- [ ] This PR will change existing database contents.
- [ ] This PR introduces a breaking change to existing installations.

## Summary

Creates resized images to facilitate faster loading

## Description of changes

This creates a resized image of images to display on navigational pages (models, library, model files, etc)
Images are cached to tmp folder, I don't know if there is a better way of going about this but the change was pretty impactful for me. Maybe worth generating previews on discovery and attaching them to corresponding file in db?

Here are some before and after load times, loading a page with 24 items displayed

| Test# | Page 1 Before |  Page 1 After | Page 2 Before |  Page 2 After |
| ------------- | ------------- | ------------- | ------------- | ------------- |
| 1 | 5.09s | .863s | 12.43s | .794s |
| 2 | 8.94s | 1.25s | 9.79s | .775s |
| 3 | 3.74s | .762s | 8.67s | 1s |
| 4 | 5.73s |  .771s | 8.05s | 1.12s |

Let me know if you want me to pr this in earnest