@Floppy 2024-02-19
Very nice, I like it! Some of those images can be really heavy, so this is a great idea.

Apologies I'm not getting through merging your PRs very quickly - my wife broke her arm, so I'm hardly getting any work time at the moment! Bear with me for a bit, I'll get there 😄