@Floppy 2024-06-21
FYI, I'm working on a rewrite of the file backend using Shrine, which has built-in support for image resizing, so I suspect we won't need this in its current form. Sorry!